+++ Mlem/Views/Tabs/Posts/Components/Post/Components/Community View/Components/Sidebar Header.swift
@@ -1,0 +1,64 @@
+//
+//  Sidebar Header.swift
+//  Mlem
+//
+//  Created by Jake Shirley on 6/21/23.
+//
+
+import SwiftUI
+import CachedAsyncImage
+
+struct CommunitySidebarHeader : View {
+    @State var communityDetails: GetCommunityResponse
+    
+    private func getRelativeTime(date: Date) -> String {
+        let formatter = RelativeDateTimeFormatter()
+        formatter.unitsStyle = .full
+        
+        return formatter.localizedString(for: date, relativeTo: Date.now)
+    }
+    
+    var body: some View {
+        ZStack(alignment: .top) {
+            // Banner
+            VStack {
+                if let bannerUrl = communityDetails.communityView.community.banner {
+                    CachedAsyncImage(url: bannerUrl) { image in
+                        image.centerCropped()
+                    } placeholder: {
+                        ProgressView()
+                    }.frame(height: 200)
+                }
+            }
+            VStack {
+                HStack(alignment: .top) {
+                    VStack(alignment: .leading) {
+                        Spacer().frame(height: 110)
+                        CommunitySidebarHeaderAvatar(imageUrl: communityDetails.communityView.community.icon)
+                        
+                        HStack {
+                            Text("Created \(getRelativeTime(date: communityDetails.communityView.community.published))")
+                        }.foregroundColor(.gray)
+                        
+                    }.padding([.leading])
+                    
+                    Spacer()
+                    
+                    VStack(alignment: .trailing) {
+                        Spacer().frame(height: 170)
+                        HStack {
+                            CommunitySidebarHeaderLabel("\(communityDetails.communityView.counts.subscribers) Subscribers")
+                            CommunitySidebarHeaderLabel("\(communityDetails.online) Online")
+                        }
+                        Spacer().frame(height: 20)
+                        
+                        Text(communityDetails.communityView.community.name).font(.title).bold().lineLimit(1)
+                        Text("@\(communityDetails.communityView.community.name)@\(communityDetails.communityView.community.actorId.host()!)").font(.footnote).lineLimit(1)
+                        
+                    }.padding([.trailing])
+                }
+            }
+        }
+    }
+}
+
