+++ Mlem/Views/Tabs/Posts/Components/Post/Components/Community View/Components/Sidebar.swift
@@ -6,102 +6,6 @@
 //
 
 import SwiftUI
-import CachedAsyncImage
-
-struct CommunitySidebarHeaderLabel : View {
-    @State var text: String
-    
-    init(_ text: String) {
-        self._text = State(initialValue: text)
-    }
-    
-    var body: some View {
-        Text(text).padding(3).foregroundColor(.white).background(RoundedRectangle(cornerRadius: 5).foregroundColor(.gray)).font(.footnote)
-    }
-}
-
-struct CommunitySidebarHeaderAvatarView : View {
-    @State var imageUrl: URL?
-    
-    var body: some View {
-        ZStack {
-            if let avatarURL = imageUrl {
-                CachedAsyncImage(url: avatarURL) { image in
-                    image
-                        .resizable()
-                        .clipShape(Circle())
-                        .overlay(Circle()
-                        .stroke(.secondary, lineWidth: 2))
-                } placeholder: {
-                    ProgressView()
-                }
-            } else {
-                // TODO: Default avatar?
-                // TIL you cannot fill and stroke at the same time
-                Circle().strokeBorder(.background, lineWidth: 2)
-                    .background(Circle().fill(.secondary))
-            }
-        }
-        .frame(width: 120, height: 120)
-        .shadow(radius: 10)
-        .background(Circle()
-        .foregroundColor(.systemBackground))
-    }
-}
-
-struct CommunitySidebarHeader : View {
-    @State var communityDetails: GetCommunityResponse
-    
-    private func getRelativeTime(date: Date) -> String {
-        let formatter = RelativeDateTimeFormatter()
-        formatter.unitsStyle = .full
-        
-        return formatter.localizedString(for: date, relativeTo: Date.now)
-    }
-    
-    var body: some View {
-        ZStack(alignment: .top) {
-            // Banner
-            VStack {
-                if let bannerUrl = communityDetails.communityView.community.banner {
-                    CachedAsyncImage(url: bannerUrl) { image in
-                        image.centerCropped()
-                    } placeholder: {
-                        ProgressView()
-                    }.frame(height: 200)
-                }
-            }
-            VStack {
-                HStack(alignment: .top) {
-                    VStack(alignment: .leading) {
-                        Spacer().frame(height: 110)
-                        CommunitySidebarHeaderAvatarView(imageUrl: communityDetails.communityView.community.icon)
-                        
-                        HStack {
-                            Text("Created \(getRelativeTime(date: communityDetails.communityView.community.published))")
-                        }.foregroundColor(.gray)
-                        
-                    }.padding([.leading])
-                    
-                    Spacer()
-                    
-                    VStack(alignment: .trailing) {
-                        Spacer().frame(height: 170)
-                        HStack {
-                            CommunitySidebarHeaderLabel("\(communityDetails.communityView.counts.subscribers) Subscribers")
-                            CommunitySidebarHeaderLabel("\(communityDetails.online) Online")
-                        }
-                        Spacer().frame(height: 20)
-                        
-                        Text(communityDetails.communityView.community.name).font(.title).bold().lineLimit(1)
-                        Text("@\(communityDetails.communityView.community.name)@\(communityDetails.communityView.community.actorId.host()!)").font(.footnote).lineLimit(1)
-                        
-                    }.padding([.trailing])
-                }
-            }
-        }
-    }
-}
 
 struct CommunitySidebarView: View {
